django 1.7 goodness, register signals
