+++ spirit/apps.py
@@ -9,3 +9,16 @@ class SpiritConfig(AppConfig):
 
     name = 'spirit'
     verbose_name = "Spirit"
+
+    def ready(self):
+        self.register_djconfig()
+        self.register_signals()
+
+    def register_djconfig(self):
+        import djconfig
+        from spirit.forms.admin import BasicConfigForm
+
+        djconfig.register(BasicConfigForm)
+
+    def register_signals(self):
+        from spirit.signals import handlers
