+++ spirit/models/comment_bookmark.py
@@ -5,10 +5,8 @@ from __future__ import unicode_literals
 from django.db import models
 from django.utils.translation import ugettext_lazy as _
 from django.conf import settings
-from django.db import IntegrityError
 from django.utils.encoding import python_2_unicode_compatible
 
-from spirit.signals.topic import topic_viewed
 from ..utils import paginator
 
 
@@ -35,24 +33,3 @@ class CommentBookmark(models.Model):
         return "%s bookmarked comment %s in %s" % (self.user.username,
                                                    self.topic.title,
                                                    self.comment_number)
-
-
-def topic_page_viewed_handler(sender, request, topic, **kwargs):
-    if not request.user.is_authenticated():
-        return
-
-    try:
-        page_number = int(request.GET.get(settings.ST_COMMENTS_PAGE_VAR, 1))
-    except ValueError:
-        return
-
-    comment_number = settings.ST_COMMENTS_PER_PAGE * (page_number - 1) + 1
-
-    try:
-        CommentBookmark.objects.update_or_create(user=request.user, topic=topic,
-                                                 defaults={'comment_number': comment_number, })
-    except IntegrityError:
-        pass
-
-
-topic_viewed.connect(topic_page_viewed_handler, dispatch_uid=__name__)
