+++ spirit/models/topic_unread.py
@@ -5,14 +5,9 @@ from __future__ import unicode_literals
 from django.db import models
 from django.utils.translation import ugettext_lazy as _
 from django.conf import settings
-from django.utils import timezone
-from django.db import IntegrityError
 from django.utils.encoding import python_2_unicode_compatible
 
-from spirit.signals.comment import comment_posted
-
 from spirit.managers.topic_unread import TopicUnreadManager
-from spirit.signals.topic import topic_viewed
 
 
 @python_2_unicode_compatible
@@ -37,24 +32,3 @@ class TopicUnread(models.Model):
 
     def __str__(self):
         return "%s read %s" % (self.user, self.topic)
-
-
-def topic_page_viewed_handler(sender, request, topic, **kwargs):
-    if not request.user.is_authenticated():
-        return
-
-    try:
-        TopicUnread.objects.update_or_create(user=request.user, topic=topic,
-                                             defaults={'is_read': True, })
-    except IntegrityError:
-        pass
-
-
-def comment_posted_handler(sender, comment, **kwargs):
-    TopicUnread.objects.filter(topic=comment.topic)\
-        .exclude(user=comment.user)\
-        .update(is_read=False, date=timezone.now())
-
-
-topic_viewed.connect(topic_page_viewed_handler, dispatch_uid=__name__)
-comment_posted.connect(comment_posted_handler, dispatch_uid=__name__)
