+++ spirit/signals/handlers/__init__.py
@@ -1,0 +1,9 @@
+# -*- coding: utf-8 -*-
+
+from . import comment
+from . import comment_bookmark
+from . import comment_history
+from . import topic
+from . import topic_notification
+from . import topic_poll
+from . import topic_unread
